YTB-21 Refactored ytbdownloader module. Changed reports.sh
