+++ quality/reports/junit/junit.xml
@@ -1,1 +1,1 @@
-<?xml version="1.0" encoding="utf-8"?><testsuites><testsuite name="pytest" errors="0" failures="0" skipped="0" tests="2" time="0.184" timestamp="2021-11-14T01:41:45.932143" hostname="EPRUPETW0176"><testcase classname="tests.test_sample.TestSample" name="test_sum" time="0.002" /><testcase classname="tests.test_sample.TestSample" name="test_sum_tuple" time="0.002" /></testsuite></testsuites>
+<?xml version="1.0" encoding="utf-8"?><testsuites><testsuite name="pytest" errors="0" failures="0" skipped="0" tests="2" time="0.296" timestamp="2021-11-14T11:18:36.157113" hostname="EPRUPETW0176"><testcase classname="tests.test_sample.TestSample" name="test_sum" time="0.003" /><testcase classname="tests.test_sample.TestSample" name="test_sum_tuple" time="0.003" /></testsuite></testsuites>
